% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pseudogp.R
\name{cov_matrix}
\alias{cov_matrix}
\title{Calculate a double exponential covariance matrix
with a diagonal noise component}
\usage{
cov_matrix(t1, t2, lambda, sigma = NULL)
}
\description{
Calculate a double exponential covariance matrix
with a diagonal noise component
}

