% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudogp.R
\docType{data}
\name{le_fit}
\alias{le_fit}
\title{Stan fit for laplacian eigenmaps representation of monocle}
\format{An object of class \code{stanfit} of dimension 500 x 1 x 161.}
\usage{
le_fit
}
\description{
Stan fit for laplacian eigenmaps representation of monocle
}
\keyword{datasets}
