% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{makeEnvelopePlot}
\alias{makeEnvelopePlot}
\title{Make envelope plot}
\usage{
makeEnvelopePlot(pst, l, s, x, chains, posterior_mean, ncurves, nnt,
  point_colour = "darkred", curve_colour = "black", point_alpha = 1,
  curve_alpha = 0.5, use_cowplot = TRUE, standardize_ranges = FALSE)
}
\arguments{
\item{pst}{Pseudotime param}

\item{l}{Lambda param}

\item{s}{Sigma param}

\item{x}{Dimred param}

\item{chains}{Number of chains}

\item{posterior_mean}{Logical. If TRUE (default) then the posterior mean curve is
plotting at the MAP estimates of all inferred parameters. If FALSE, then \code{nsamples}
will be randomly drawn from the posterior of the parameters and a curve plotted for each.}

\item{ncurves}{The number of new points at which to calculate the posterior mean curves.}

\item{nnt}{Number of new pseudo time points at which to plot the posterior mean curves.}

\item{point_colour}{The colour of the points (cells) to draw}

\item{curve_colour}{The colour of the curves to draw}

\item{point_alpha}{The alpha (opacity) of the points}

\item{curve_alpha}{The alpha (opacity) of the curves. Note that this is just a suggested value
and the function will choose an appropriate value depending on the number of samples to plot.
This is chosen as \eqn{(1 - \alpha) * exp(1 - nsamples) + \alpha}.}

\item{use_cowplot}{Logical. Determines whether to use \code{\link[cowplot]{theme_cowplot}} or \code{\link[ggplot2]{theme_bw}}.}

\item{standardize_ranges}{Logical. If plotting multiple representations it can be useful to have
x and y lims that don't depend on the fit (so plots align correctly). If this is set to FALSE,
\code{ggplot2} calculates the x and y limits. If this is set to TRUE, the x and y limits are set
to the minimum and maximum of the X values plus or minus 6\% of the range between them.}
}
\description{
Make envelope plot
}
