% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudogp.R
\name{meanvar_to_alphabeta}
\alias{meanvar_to_alphabeta}
\title{Convert the mean and variance of a gamma distribution
to the alpha-beta parametrization}
\usage{
meanvar_to_alphabeta(mean, var)
}
\description{
Convert the mean and variance of a gamma distribution
to the alpha-beta parametrization
}
