% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudogp.R
\docType{data}
\name{monocle_le}
\alias{monocle_le}
\title{Laplacian eigenmaps representation of monocle data}
\format{An object of class \code{matrix} with 155 rows and 2 columns.}
\usage{
monocle_le
}
\description{
The 'monocle' data comes from Trapnell et al. (2014) (\url{})
}
\keyword{datasets}

