% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudogp.R
\docType{data}
\name{monocle_pca}
\alias{monocle_pca}
\title{PCA representation of monocle data}
\format{An object of class \code{matrix} with 155 rows and 2 columns.}
\usage{
monocle_pca
}
\description{
PCA representation of monocle data
}
\keyword{datasets}

