% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pseudogp.R
\docType{data}
\name{monocle_tsne}
\alias{monocle_tsne}
\title{tSNE representation of monocle data}
\format{\preformatted{ num [1:155, 1:2] -1.07 -0.93 -0.64 -1.58 0.75 ...
 - attr(*, "dimnames")=List of 2
  ..$ : chr [1:155] "T0_CT_A01" "T0_CT_A03" "T0_CT_A05" "T0_CT_A06" ...
  ..$ : NULL
}}
\usage{
monocle_tsne
}
\description{
tSNE representation of monocle data
}
\keyword{datasets}

