% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pseudogp.R
\name{posterior_mean_curve}
\alias{posterior_mean_curve}
\title{Compute the posterior mean curve}
\usage{
posterior_mean_curve(X, t, l, s, nnt = 80)
}
\arguments{
\item{X}{The data representation}

\item{t}{The pseudotime sample}

\item{l}{Lambda}

\item{s}{Signa}

\item{nnt}{Number of new time points}
}
\description{
This computes the posterior mean curve. Note that the "mean" aspect here
doesn't relate to pseudotimes - this can take a pseudotime sample from the
posterior pseudotime distribution and will calculate the mean curve
in the data space.
}

