% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{posteriorBoxplot}
\alias{posteriorBoxplot}
\title{Posterior boxplot ordered by pseudotime}
\usage{
posteriorBoxplot(fit, inner = 0.75, outer = 0.95)
}
\arguments{
\item{fit}{The fit object returned by \code{fitPseudotime}}

\item{inner}{The inner interval for box 'edges' (default 0.75)}

\item{outer}{The outer interval for boxplot whiskers (default 0.95)}
}
\value{
A \code{ggplot2} boxplot of posterior pseudotime samples ordered by
median pseudotime.
}
\description{
Posterior boxplot ordered by pseudotime
}

