% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pseudogp.R
\name{fitPseudotime}
\alias{fitPseudotime}
\title{Fit probabilistic pseudotime model}
\usage{
fitPseudotime(X, smoothing_alpha = 10, smoothing_beta = 3,
  pseudotime_mean = 0.5, pseudotime_var = 1, chains = 1, iter = 1000,
  ...)
}
\description{
Fit probabilistic pseudotime model
}

