% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pseudogp.R
\docType{data}
\name{monocle_le}
\alias{monocle_le}
\title{Laplacian eigenmaps representation of monocle data}
\format{\preformatted{ num [1:155, 1:2] -1.152 -0.508 -0.382 -1.009 -0.558 ...
}}
\usage{
monocle_le
}
\description{
Laplacian eigenmaps representation of monocle data
}
\keyword{datasets}

