% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pseudogp.R
\docType{data}
\name{monocle_pca}
\alias{monocle_pca}
\title{PCA representation of monocle data}
\format{\preformatted{ num [1:155, 1:2] -1.47 -1.7 -1.42 -1.39 -1.8 ...
 - attr(*, "dimnames")=List of 2
  ..$ : chr [1:155] "T0_CT_A01" "T0_CT_A03" "T0_CT_A05" "T0_CT_A06" ...
  ..$ : chr [1:2] "PC1" "PC2"
}}
\usage{
monocle_pca
}
\description{
PCA representation of monocle data
}
\keyword{datasets}

