% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{posteriorCurvePlot}
\alias{posteriorCurvePlot}
\title{Posterior curve plot (workhorse of the analysis)}
\usage{
posteriorCurvePlot(X, fit, nsamples = 50, nnt = 80, posterior_mean = TRUE,
  ...)
}
\description{
Posterior curve plot (workhorse of the analysis)
}

