% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudogp.R
\name{standardize}
\alias{standardize}
\title{Column-wise standardize input data to mean 0 and variance 1}
\usage{
standardize(X)
}
\arguments{
\item{X}{data to standardize}
}
\description{
Column-wise standardize input data to mean 0 and variance 1
}
